+++ lib/mix/tasks/ua_inspector/download.ex
@@ -36,7 +36,7 @@ defmodule Mix.Tasks.UaInspector.Download do
   ]
 
   def run(args) do
-    :ok = start_app()
+    :ok = load_app(args)
     :ok = Config.init_env()
 
     {opts, _argv, _errors} = OptionParser.parse(args, @cli_options)
@@ -80,16 +80,11 @@ defmodule Mix.Tasks.UaInspector.Download do
     end
   end
 
-  defp start_app do
-    project = Mix.Project.config()
+  defp load_app(args) do
+    _ = Mix.Task.run("loadpaths", args)
 
-    if project[:app] not in Application.started_applications() do
-      startup_silent = Application.get_env(:ua_inspector, :startup_silent, false)
-      :ok = Application.put_env(:ua_inspector, :startup_silent, true)
-
-      _ = Mix.Task.run("app.start")
-
-      :ok = Application.put_env(:ua_inspector, :startup_silent, startup_silent)
+    unless "--no-compile" in args do
+      _ = Mix.Task.run("compile", args)
     end
 
     :ok